added reverse and r_reverse

+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: lruiz-to <lruiz-to@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/25 14:24:55 by lruiz-to          #+#    #+#             */
-/*   Updated: 2025/03/25 14:27:08 by lruiz-to         ###   ########.fr       */
+/*   Updated: 2025/04/10 14:31:32 by lruiz-to         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 long	ft_atol(char *num)
 {
-	long	res;
-	long	sign;
-	int		cont;
+	long long int	res;
+	long			sign;
+	int				cont;
 
 	res = 0;
 	cont = 0;
